backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
boto3==1.35.24
PyJWT[crypto]>=2.10.1,<3.0.0
python-multipart==0.0.9
email-validator==2.2.0
